pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemini-mcp-connect"
version = "1.0.0"
description = "Claude Code plugin: Google Gemini as a second-opinion AI via MCP"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["claude", "mcp", "gemini", "ai", "code-review"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "platformdirs>=4.0.0",
]

[project.scripts]
gemini-mcp-connect     = "gemini_bridge.server:main"
gemini-mcp-connect-cli = "gemini_bridge.cli:main"

[project.urls]
Homepage   = "https://github.com/StealthyLabsHQ/gemini-mcp-connect"
Repository = "https://github.com/StealthyLabsHQ/gemini-mcp-connect"
Issues     = "https://github.com/StealthyLabsHQ/gemini-mcp-connect/issues"

[tool.hatch.build.targets.wheel]
packages = ["gemini_bridge"]
